requirement.txt
selenium==3.0
pytesseract
opencv-pyth
scikit-learn
numpy
matplotlib
ddddocr
requests
Pillow==9.5.0